refactor(pdf): restructure pdf directories to improve modularity and readability
