+++ lib/pages/create_a_character_pages/finishing_up_tab.dart
@@ -1,7 +1,7 @@
 import "package:flutter/material.dart";
 
 import "../../content_classes/all_content_classes.dart";
-import "../../pdf/pdf_display.dart";
+import "../../pdf/display.dart";
 import "../../storage/global_list_manager.dart";
 import "../../theme/theme_manager.dart";
 import "../../utils/style_utils.dart";
