+++ lib/pages/my_characters_pages/my_characters.dart
@@ -2,7 +2,7 @@
 import "package:flutter/material.dart";
 
 import "../../content_classes/all_content_classes.dart";
-import "../../pdf/pdf_display.dart";
+import "../../pdf/display.dart";
 import "../../storage/global_list_manager.dart";
 import "../../theme/theme_manager.dart";
 import "../../utils/style_utils.dart";
