+++ lib/pdf/first_page/second_column.dart
@@ -1,0 +1,9 @@
+// External Imports
+import "package:pdf/widgets.dart";
+
+// Project Imports
+import "../../content_classes/all_content_classes.dart";
+
+Widget buildSecondColumn(Character userCharacter) {
+  return Container();
+}
