+++ lib/pdf/first_page/third_column.dart
@@ -1,0 +1,9 @@
+// External Imports
+import "package:pdf/widgets.dart";
+
+// Project Imports
+import "../../content_classes/all_content_classes.dart";
+
+Widget buildThirdColumn(Character userCharacter) {
+  return Container();
+}
